+++ src/index.ts
@@ -25,6 +25,7 @@ export {
   SortMeta,
   FilterMeta,
   SkipGlobalErrors,
+  Deprecated,
 } from './decorators';
 
 // Interfaces
@@ -49,6 +50,10 @@ export type {
   SortInfo,
   SwaggerOptions,
   ContextOptions,
+  DeprecatedOptions,
+  DeprecationMeta,
+  RateLimitOptions,
+  RateLimitMeta,
 } from './interfaces';
 
 // Adapters
@@ -67,10 +72,12 @@ export {
   ProblemDetailsDto,
   SortMetaDto,
   FilterMetaDto,
+  DeprecationMetaDto,
+  RateLimitMetaDto,
 } from './dto/response.dto';
 
 // Constants
-export { DEFAULT_ERROR_CODE_MAP, DEFAULT_PROBLEM_TITLE_MAP, CURSOR_PAGINATED_KEY, PROBLEM_TYPE_KEY, SORT_META_KEY, FILTER_META_KEY, SKIP_GLOBAL_ERRORS_KEY } from './constants';
+export { DEFAULT_ERROR_CODE_MAP, DEFAULT_PROBLEM_TITLE_MAP, CURSOR_PAGINATED_KEY, PROBLEM_TYPE_KEY, SORT_META_KEY, FILTER_META_KEY, SKIP_GLOBAL_ERRORS_KEY, DEPRECATED_KEY } from './constants';
 
 // Swagger utilities
 export { applyGlobalErrors } from './swagger/global-errors';